breaks at tbb download

+++ LoLUpdater/LoLUpdater.cpp
@@ -29,50 +29,65 @@ const std::wstring g_szClassName(L"mainwindow1");
 RECT start = { 0, 0, 100, 20 };
 RECT end = { 0, 100, 100, 120 };
 
-void download(std::wstring fromurl, std::wstring dest, std::wstring args)
+void download(const std::wstring url, const std::wstring file, const std::wstring args)
 {
-	URLDownloadToFile(
+	HRESULT hRez = URLDownloadToFile(
 		nullptr,
-		fromurl.c_str(),
-		dest.c_str(),
+		url.c_str(),
+		file.c_str(),
 		0,
 		nullptr
 		);
-	std::wstring *unblocker = nullptr;
-	PathCombine(reinterpret_cast<LPWSTR>(&unblocker), reinterpret_cast<LPWSTR>(&cwd), dest.c_str());
-	pathcontainer[0] << (reinterpret_cast<LPWSTR>(&unblocker) + constants[0]);
-	DeleteFile(pathcontainer[0].str().c_str());
-	pathcontainer[0].str(std::wstring());
-	pathcontainer[0].clear();
-	SHELLEXECUTEINFO ShExecInfo = { 0 };
-	ShExecInfo.cbSize = sizeof(SHELLEXECUTEINFO);
-	ShExecInfo.fMask = NULL;
-	ShExecInfo.hwnd = nullptr;
-	ShExecInfo.lpVerb = nullptr;
-	ShExecInfo.lpFile = dest.c_str();
-	ShExecInfo.lpParameters = args.c_str();
-	ShExecInfo.lpDirectory = nullptr;
-	ShExecInfo.nShow = SW_HIDE;
-	ShExecInfo.hInstApp = nullptr;
-	ShellExecuteEx(&ShExecInfo);
-	WaitForSingleObject(
-		ShExecInfo.hProcess,
-		INFINITE
-		);
+
+	if (hRez == 0)
+	{
+		std::wstring *unblocker = nullptr;
+		PathCombine(reinterpret_cast<LPWSTR>(unblocker), reinterpret_cast<LPWSTR>(&cwd[0]), file.c_str());
+		pathcontainer[0] << (std::wstring(reinterpret_cast<LPWSTR>(&unblocker)) + constants[0]);
+		DeleteFile(pathcontainer[0].str().c_str());
+		pathcontainer[0].str(std::wstring());
+		pathcontainer[0].clear();
+
+		SHELLEXECUTEINFO ShExecInfo = { 0 };
+		ShExecInfo.cbSize = sizeof(SHELLEXECUTEINFO);
+		ShExecInfo.fMask = NULL;
+		ShExecInfo.hwnd = nullptr;
+		ShExecInfo.lpVerb = nullptr;
+		ShExecInfo.lpFile = file.c_str();
+		ShExecInfo.lpParameters = args.c_str();
+		ShExecInfo.lpDirectory = nullptr;
+		ShExecInfo.nShow = SW_SHOW;
+		ShExecInfo.hInstApp = nullptr;
+		ShellExecuteEx(&ShExecInfo);
+		WaitForSingleObject(
+			ShExecInfo.hProcess,
+			INFINITE
+			);
+	}
+	
 }
 
 void tbbdownload(const std::wstring file)
 {
-	const std::wstring ftp(L"http://lol.jdhpro.com/");
-	pathcontainer[0] << (ftp + file.c_str());
+	std::wstring *ftp[MAX_PATH+1];
+
+	UrlCombine(
+		L"http://lol.jdhpro.com/",
+		file.c_str(),
+		reinterpret_cast<LPWSTR>(ftp),
+		reinterpret_cast<LPDWORD>(MAX_PATH + 1),
+		0
+		);
+
 
 	URLDownloadToFile(
 		nullptr,
-		pathcontainer[0].str().c_str(),
+		reinterpret_cast<LPWSTR>(ftp),
 		reinterpret_cast<LPWSTR>(tbb),
 		0,
 		nullptr
 		);
+
 }
 
 LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam)
@@ -136,34 +151,34 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	AllocConsole();
 	_wfreopen(L"CONOUT$", L"w", stdout);
 #endif
-	std::wstring buffer_1[MAX_PATH] = { L"projects" };
+	std::wstring buffer_1[MAX_PATH + 1] = { L"projects" };
 	std::wstring *lpStr1;
 	lpStr1 = buffer_1;
 	PathAppend(reinterpret_cast<LPWSTR>(lpStr1), L"lol_air_client");
 #ifdef DEBUG
-	wprintf(reinterpret_cast<LPWSTR>(buffer_1));
+	wprintf(reinterpret_cast<LPWSTR>(lpStr1));
 #endif
 
-	std::wstring buffer_3[MAX_PATH] = { L"solutions" };
+	std::wstring buffer_3[MAX_PATH + 1] = { L"solutions" };
 	std::wstring *lpStr3;
 	lpStr3 = buffer_3;
 	PathAppend(reinterpret_cast<LPWSTR>(lpStr3), L"lol_game_client_sln");
 #ifdef DEBUG
-	wprintf(reinterpret_cast<LPWSTR>(buffer_3));
+	wprintf(reinterpret_cast<LPWSTR>(lpStr3));
 #endif
 
 	const std::wstring rads(L"RADS");
 	// air
-	std::wstring buffer_2[MAX_PATH] = { rads };
+	std::wstring buffer_2[MAX_PATH + 1] = { rads };
 	std::wstring *lpStr2;
 	lpStr2 = buffer_2;
 
-	PathAppend(reinterpret_cast<LPWSTR>(lpStr2), reinterpret_cast<LPWSTR>(buffer_1));
+	PathAppend(reinterpret_cast<LPWSTR>(lpStr2), reinterpret_cast<LPWSTR>(lpStr1));
 #ifdef DEBUG
-	wprintf(reinterpret_cast<LPWSTR>(buffer_2));
+	wprintf(reinterpret_cast<LPWSTR>(lpStr2));
 #endif
 	// game
-	std::wstring buffer_4[MAX_PATH] = { rads };
+	std::wstring buffer_4[MAX_PATH + 1] = { rads };
 	std::wstring *lpStr4;
 	lpStr4 = buffer_4;
 	PathAppend(reinterpret_cast<LPWSTR>(lpStr4), reinterpret_cast<LPWSTR>(lpStr3));
@@ -172,33 +187,33 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #endif
 
 	const std::wstring rel(L"releases");
-	std::wstring buffer_5[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(buffer_4)) };
+	std::wstring buffer_5[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(lpStr4)) };
 	std::wstring *gameclient;
 	gameclient = buffer_5;
 	PathAppend(reinterpret_cast<LPWSTR>(gameclient), rel.c_str());
 #ifdef DEBUG
 	wprintf(reinterpret_cast<LPWSTR>(buffer_5));
 #endif
-	std::wstring buffer_6[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(lpStr2)) };
+	std::wstring buffer_6[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(lpStr2)) };
 	std::wstring *airclient;
 	airclient = buffer_6;
 	PathAppend(reinterpret_cast<LPWSTR>(airclient), rel.c_str());
 #ifdef DEBUG
-	wprintf(reinterpret_cast<LPWSTR>(buffer_6));
+	wprintf(reinterpret_cast<LPWSTR>(airclient));
 #endif
 
 
 	download(L"http://developer.download.nvidia.com/cg/Cg_3.1/Cg-3.1_April2012_Setup.exe", L"Cg-3.1_April2012_Setup.exe", L"/verysilent /TYPE = compact");
-	std::vector<std::wstring> cgbinpath(MAX_PATH, std::wstring());
+	std::vector<std::wstring> cgbinpath(MAX_PATH + 1, std::wstring());
 	GetEnvironmentVariable(L"CG_BIN_PATH",
 		reinterpret_cast<LPWSTR>(&cgbinpath[0]),
-		MAX_PATH);
+		MAX_PATH + 1);
 #ifdef DEBUG
 	wprintf(reinterpret_cast<LPWSTR>(&cgbinpath[0]));
 #endif
 	wcsncat_s(
 		reinterpret_cast<LPWSTR>(&cgbinpath[0]),
-		MAX_PATH,
+		MAX_PATH + 1,
 		L"\\",
 		_TRUNCATE
 		);
@@ -207,11 +222,11 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #endif
 
 	std::wstring progfiles = std::wstring(L":") + std::wstring(L"Program Files");
-	pathcontainer[1] << std::wstring(reinterpret_cast<LPWSTR>(&cwd))[0] + progfiles;
+	pathcontainer[1] << std::wstring(reinterpret_cast<LPWSTR>(&cwd[0]))[0] + progfiles;
 #ifdef DEBUG
 	wprintf(pathcontainer[1].str().c_str());
 #endif
-	if (sizeof(void*) == 4)
+	if (sizeof(void*) == 8)
 	{
 		pathcontainer[1] << std::wstring(L" (x86)");
 #ifdef DEBUG
@@ -221,11 +236,11 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 
 	download(L"https://labsdownload.adobe.com/pub/labs/flashruntimes/air/air15_win.exe", L"air15_win.exe", L"-silent");
 
-	pathcontainer[1] << std::wstring(L"Common Files") + std::wstring(L"Adobe ") + constants[2];
+	pathcontainer[1] << std::wstring(L"\\Common Files") + std::wstring(L"\\Adobe ") + constants[2];
 #ifdef DEBUG
 	wprintf(pathcontainer[1].str().c_str());
 #endif
-	std::wstring adobepath[MAX_PATH];
+	std::wstring adobepath[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(adobepath),
 		pathcontainer[1].str().c_str(),
@@ -237,7 +252,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	
 
 	const std::wstring cg(L"Cg.dll");
-	std::wstring cgbin[MAX_PATH];
+	std::wstring cgbin[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cgbin),
 		reinterpret_cast<LPWSTR>(&cgbinpath[0]),
@@ -247,7 +262,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(cgbin));
 #endif
 	const std::wstring cggl(L"CgGL.dll");
-	std::wstring cgglbin[MAX_PATH];
+	std::wstring cgglbin[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cgglbin),
 		reinterpret_cast<LPWSTR>(&cgbinpath[0]),
@@ -257,7 +272,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(cgglbin));
 #endif
 	const std::wstring cgd3d9(L"CgD3D9.dll");
-	std::wstring cgd3d9bin[MAX_PATH];
+	std::wstring cgd3d9bin[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cgd3d9bin),
 		reinterpret_cast<LPWSTR>(&cgbinpath[0]),
@@ -267,7 +282,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(cgd3d9bin));
 #endif
 
-	std::wstring buffer_7[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(buffer_5)) };
+	std::wstring buffer_7[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(gameclient)) };
 	std::wstring *random7;
 	random7 = buffer_7;
 	PathAppend(reinterpret_cast<LPWSTR>(random7), L"0.0.1.64");
@@ -275,7 +290,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(random7));
 #endif
 	const std::wstring dep(L"deploy");
-	std::wstring buffer_12[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(random7)) };
+	std::wstring buffer_12[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(random7)) };
 	std::wstring *version;
 	version = buffer_12;
 	PathAppend(reinterpret_cast<LPWSTR>(version), dep.c_str());
@@ -285,14 +300,14 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 
 
 
-	std::wstring buffer_14[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(buffer_6)) };
+	std::wstring buffer_14[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(airclient)) };
 	std::wstring*lpStr14;
 	lpStr14 = buffer_14;
 	PathAppend(reinterpret_cast<LPWSTR>(lpStr14), L"0.0.1.117");
 #ifdef DEBUG
 	wprintf(reinterpret_cast<LPWSTR>(buffer_14));
 #endif
-	std::wstring buffer_13[MAX_PATH] = { std::wstring(reinterpret_cast<LPWSTR>(buffer_14)) };
+	std::wstring buffer_13[MAX_PATH + 1] = { std::wstring(reinterpret_cast<LPWSTR>(lpStr14)) };
 	std::wstring *lpStr13;
 	lpStr13 = buffer_13;
 	PathAppend(reinterpret_cast<LPWSTR>(lpStr13), dep.c_str());
@@ -301,7 +316,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #endif
 
 
-	std::wstring buffer_15[MAX_PATH] = { constants[1].c_str() };
+	std::wstring buffer_15[MAX_PATH + 1] = { constants[1].c_str() };
 	std::wstring *airversion;
 	airversion = buffer_15;
 	PathAppend(reinterpret_cast<LPWSTR>(airversion), reinterpret_cast<LPWSTR>(lpStr13));
@@ -329,7 +344,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #ifdef DEBUG
 		wprintf(reinterpret_cast<LPWSTR>(version));
 #endif
-		std::wstring buffer_17[MAX_PATH] = { constants[1].c_str() };
+		std::wstring buffer_17[MAX_PATH + 1] = { constants[1].c_str() };
 		std::wstring *airversion;
 		airversion = buffer_17;
 		PathAppend(reinterpret_cast<LPWSTR>(airversion), constants[2].c_str());
@@ -338,11 +353,11 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #endif
 	}
 
-	pathcontainer[2] << (std::wstring(reinterpret_cast<LPWSTR>(&cwd)) + std::wstring(reinterpret_cast<LPWSTR>(buffer_12)));
+	pathcontainer[2] << (std::wstring(reinterpret_cast<LPWSTR>(&cwd[0])) + std::wstring(reinterpret_cast<LPWSTR>(buffer_12)));
 #ifdef DEBUG
 	wprintf(pathcontainer[2].str().c_str());
 #endif
-	pathcontainer[3] << (std::wstring(reinterpret_cast<LPWSTR>(&cwd)) + std::wstring(reinterpret_cast<LPWSTR>(buffer_15)));
+	pathcontainer[3] << (std::wstring(reinterpret_cast<LPWSTR>(&cwd[0])) + std::wstring(reinterpret_cast<LPWSTR>(buffer_15)));
 #ifdef DEBUG
 	wprintf(pathcontainer[3].str().c_str());
 #endif
@@ -356,7 +371,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(L"tbb path", reinterpret_cast<LPWSTR>(tbb));
 #endif
 	const std::wstring air(L"Adobe AIR.dll");
-	std::wstring airdest[MAX_PATH];
+	std::wstring airdest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(airdest),
 		pathcontainer[3].str().c_str(),
@@ -365,7 +380,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #ifdef DEBUG
 	wprintf(reinterpret_cast<LPWSTR>(airdest));
 #endif
-	std::wstring airlatest[MAX_PATH];
+	std::wstring airlatest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(airlatest),
 		reinterpret_cast<LPWSTR>(adobepath),
@@ -375,7 +390,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(airlatest));
 #endif
 
-	std::wstring buffer_11[MAX_PATH] = { L"Resources" };
+	std::wstring buffer_11[MAX_PATH + 1] = { L"Resources" };
 	std::wstring *flash;
 	flash = buffer_11;
 
@@ -384,7 +399,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(flash));
 #endif
 
-	std::wstring flashdest[MAX_PATH];
+	std::wstring flashdest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(flashdest),
 		pathcontainer[3].str().c_str(),
@@ -394,7 +409,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(flashdest));
 #endif
 
-	std::wstring flashlatest[MAX_PATH];
+	std::wstring flashlatest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(flashlatest),
 		reinterpret_cast<LPWSTR>(adobepath),
@@ -404,7 +419,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(flashlatest));
 #endif
 
-	std::wstring cgdest[MAX_PATH];
+	std::wstring cgdest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cgdest),
 		pathcontainer[2].str().c_str(),
@@ -414,7 +429,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 	wprintf(reinterpret_cast<LPWSTR>(cgdest));
 #endif
 
-	std::wstring cggldest[MAX_PATH];
+	std::wstring cggldest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cggldest),
 		pathcontainer[2].str().c_str(),
@@ -423,7 +438,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
 #ifdef DEBUG
 	wprintf(reinterpret_cast<LPWSTR>(cggldest));
 #endif
-	std::wstring cgd3d9dest[MAX_PATH];
+	std::wstring cgd3d9dest[MAX_PATH + 1];
 	PathCombine(
 		reinterpret_cast<LPWSTR>(cgd3d9dest),
 		pathcontainer[2].str().c_str(),
